+++ NEWS.md
@@ -7,6 +7,7 @@
   - Improved `find_fixpoint` function with better input validation and time handling support
   - Add equality (`==`) and approximate equality (`isapprox`) methods for `NWState` and `NWParameter`
   - Minor documentation fixes and spelling corrections
+  - fix [#310](https://github.com/JuliaDynamics/NetworkDynamics.jl/pull/310) (allow `VPIndex(i)` to index into network objects)
 
 ## v0.10.4 Changelog
 - [#303](https://github.com/JuliaDynamics/NetworkDynamics.jl/pull/303) update for ModelingToolkit.jl v10 compatibility: